+++ CHANGELOG.md
@@ -29,8 +29,8 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 ### Versions
 
-- Electron 26.0.0 (was 25.4.0)
-- Chromium 116.0.5845.82 (was 114.0.5735.248)
+- Electron 27.0.0-alpha.2 (was 25.4.0)
+- Chromium 118.0.5949.0 (was 114.0.5735.248)
 
 ## [10.2.0](https://github.com/Jelmerro/Vieb/compare/10.1.1...10.2.0) - 2023-08-09